+++ documentation/samples/packs/authentication/coda.ts
@@ -1,11 +1,11 @@
 import * as sdk from "@codahq/packs-sdk";
 export const pack = sdk.newPack();
 
-// Per-user authentication to the Coda API, using a token in the Authorization
-// header.
+// Per-user authentication to the Superhuman Docs API, using a token in the
+// Authorization header.
 // See https://coda.io/developers/apis/v1
 pack.setUserAuthentication({
-  type: sdk.AuthenticationType.CodaApiHeaderBearerToken,
+  type: sdk.AuthenticationType.SuperhumanDocsApiHeaderBearerToken,
 
   // Creates the token automatically when the Pack is installed.
   shouldAutoAuthSetup: true,